Req11 Edición de alianza

+++ MVC_Project.WebBackend/Models/AllianceViewModel.cs
@@ -37,9 +37,8 @@ namespace MVC_Project.WebBackend.Models
         public string name { get; set; }
 
         [Display(Name = "Aliado")]
-        public Int16 allyId { get; set; }
+        public Int64 allyId { get; set; }
         public SelectList allyList { get; set; }
-        //public MultiSelectList allyMultList { get; set; }
 
         [Display(Name = "% comisión aliado")]
         public decimal allyCommisionPercent { get; set; }
